update informes, grupos

+++ src/Template/Grupo/view.ctp
@@ -3,9 +3,7 @@
  * @var \App\View\AppView $this
  * @var \App\Model\Entity\Taller $taller
  */
-
 use App\Enums\RolesEnum;
-
 ?>
 <div class="row">
      <?php echo $this->element('menu_grupo'); ?>    
@@ -29,15 +27,14 @@ use App\Enums\RolesEnum;
                             <tbody>
                                 <?php foreach ($alumnos as $alumno): ?>
                                 <tr>
-                                    <td><?= $alumno->image == "Null" ? $this->Html->image('../images/users/avatar-1.jpg', ['height' => 40,'class' => 'rounded-circle']) : $this->Html->image($alumno->image, ['height' => 40,'class' => 'rounded-circle'])
-                                             ?></td>
-                
+                                    <td><?= $alumno->image == "Null" ? $this->Html->image('../images/users/avatar-1.jpg', ['height' => 40,'class' => 'rounded-circle']) : $this->Html->image($alumno->image, ['height' => 40,'class' => 'rounded-circle'])?></td>
                                     <td><?= h($alumno->name . " " . $alumno->surname) ?></td>
-                                   
                                     <td class="actions">
                                         <?= $this->Html->link(__('Calificar'), ['controller' => 'ProcesoAlumnos', 'action' => 'add', $alumno->id],['class'=>'btn btn-info btn-rounded waves-effect waves-light m-t-5','type' => 'button']) ?>
-                                        <?= $this->Html->link(__('Observacion'), ['controller' => 'ObservacionesAlumnos', 'action' => 'add', $alumno->id],['class'=>'btn btn-info btn-rounded waves-effect waves-light m-t-5','type' => 'button']) ?>
-                                        <?= $this->Html->link(__('Quitar'),['controller'=>'Alumnos','action' => 'deleteByAlumno',$alumno->id, $grupo->id],['class'=>'btn btn-warning btn-rounded waves-effect waves-light m-t-5','type' => 'button']) ?>
+                                        <?= $this->Html->link(__('Observación'), ['controller' => 'ObservacionesAlumnos', 'action' => 'add', $alumno->id],['class'=>'btn btn-success btn-rounded waves-effect waves-light m-t-5','type' => 'button']) ?>
+                                        <?php if($user_session['role_id'] == RolesEnum::PROFESOR_ADMIN) { ?>
+                                            <?= $this->Html->link(__('Quitar'),['controller'=>'Alumnos','action' => 'deleteByAlumno',$alumno->id, $grupo->id],['class'=>'btn btn-warning btn-rounded waves-effect waves-light m-t-5','type' => 'button']) ?>
+                                        <?php } ?>
                                     </td>
                                 </tr>
                                 <?php endforeach; ?>
@@ -46,9 +43,7 @@ use App\Enums\RolesEnum;
                                 <tr>
                                      <td></td>
                                     <td>Cantidad de alumnos</td>
-                                
-                                   
-                                    <td scope="col"><?= $this->Number->format($cantidad)  ?></td>
+                                    <td scope="col"><?= $this->Number->format($cantidad)?></td>
                                 </tr>
                             </tfoot>
                         </table>
